Revert default.qubit sample changes (uncorrelated results when sampling multiple wires), revert organization of sampling tests from qubit_device
