+++ tests/test_default_qubit.py
@@ -1437,18 +1437,15 @@ class TestTensorExpval:
 
         A = np.array([[1.02789352, 1.61296440 - 0.3498192j], [1.61296440 + 0.3498192j, 1.23920938 + 0j]])
         Identity = np.array([[1, 0],[0, 1]])
-        print('State of device')
-        print(dev._state)
+
         obs = np.kron(np.kron(Identity,Identity), A)
         res = dev.expval(qml.Hermitian(obs, wires=[2,1,0]))
 
         a = A[0, 0]
         re_b = A[0, 1].real
         d = A[1, 1]
-        print(res)
 
         expected = ((a - d) * np.cos(theta) + 2 * re_b * np.sin(theta) * np.sin(phi) + a + d) / 2
-        print(expected)
         assert np.allclose(res, expected, atol=tol, rtol=0)
 
 @pytest.mark.parametrize("theta, phi, varphi", list(zip(THETA, PHI, VARPHI)))
@@ -1553,4 +1550,142 @@ class TestTensorVar:
 
         assert np.allclose(res, expected, atol=tol, rtol=0)
 
+@pytest.mark.parametrize("theta, phi, varphi", list(zip(THETA, PHI, VARPHI)))
+class TestTensorSample:
+    """Test tensor expectation values"""
 
+    def test_paulix_pauliy(self, theta, phi, varphi, monkeypatch, tol):
+        """Test that a tensor product involving PauliX and PauliY works correctly"""
+        dev = qml.device("default.qubit", wires=3, shots=10000)
+        dev.reset()
+        dev.apply(qml.RX(theta, wires=[0]))
+        dev.apply(qml.RX(phi, wires=[1]))
+        dev.apply(qml.RX(varphi, wires=[2]))
+        dev.apply(qml.CNOT(wires=[0, 1]))
+        dev.apply(qml.CNOT(wires=[1, 2]))
+
+        with monkeypatch.context() as m:
+            m.setattr("numpy.random.choice", lambda x, y, p: (x, p))
+            s1, p = dev.sample(qml.PauliX(0) @ qml.PauliY(2))
+
+
+        print(s1)
+        print(p)
+        # s1 should only contain 1 and -1
+        assert np.allclose(s1 ** 2, 1, atol=tol, rtol=0)
+
+        mean = s1 @ p
+        expected = np.sin(theta) * np.sin(phi) * np.sin(varphi)
+        assert np.allclose(mean, expected, atol=tol, rtol=0)
+
+        var = (s1 ** 2) @ p - (s1 @ p).real ** 2
+        expected = (
+            8 * np.sin(theta) ** 2 * np.cos(2 * varphi) * np.sin(phi) ** 2
+            - np.cos(2 * (theta - phi))
+            - np.cos(2 * (theta + phi))
+            + 2 * np.cos(2 * theta)
+            + 2 * np.cos(2 * phi)
+            + 14
+        ) / 16
+        assert np.allclose(var, expected, atol=tol, rtol=0)
+
+    def test_pauliz_hadamard(self, theta, phi, varphi, monkeypatch, tol):
+        """Test that a tensor product involving PauliZ and PauliY and hadamard works correctly"""
+        dev = qml.device("default.qubit", wires=3)
+        dev.reset()
+        dev.apply(qml.RX(theta, wires=[0]))
+        dev.apply(qml.RX(phi, wires=[1]))
+        dev.apply(qml.RX(varphi, wires=[2]))
+        dev.apply(qml.CNOT(wires=[0, 1]))
+        dev.apply(qml.CNOT(wires=[1, 2]))
+
+        with monkeypatch.context() as m:
+            m.setattr("numpy.random.choice", lambda x, y, p: (x, p))
+            s1, p = dev.sample(qml.PauliZ(0) @ qml.Hadamard(1) @ qml.PauliY(2))
+
+        # s1 should only contain 1 and -1
+        assert np.allclose(s1 ** 2, 1, atol=tol, rtol=0)
+
+        mean = s1 @ p
+        expected = -(np.cos(varphi) * np.sin(phi) + np.sin(varphi) * np.cos(theta)) / np.sqrt(2)
+        assert np.allclose(mean, expected, atol=tol, rtol=0)
+
+        var = (s1 ** 2) @ p - (s1 @ p).real ** 2
+        expected = (
+            3
+            + np.cos(2 * phi) * np.cos(varphi) ** 2
+            - np.cos(2 * theta) * np.sin(varphi) ** 2
+            - 2 * np.cos(theta) * np.sin(phi) * np.sin(2 * varphi)
+        ) / 4
+        assert np.allclose(var, expected, atol=tol, rtol=0)
+
+    def test_hermitian(self, theta, phi, varphi, monkeypatch, tol):
+        """Test that a tensor product involving qml.Hermitian works correctly"""
+        dev = qml.device("default.qubit", wires=3)
+        dev.reset()
+        dev.apply(qml.RX(theta, wires=[0]))
+        dev.apply(qml.RX(phi, wires=[1]))
+        dev.apply(qml.RX(varphi, wires=[2]))
+        dev.apply(qml.CNOT(wires=[0, 1]))
+        dev.apply(qml.CNOT(wires=[1, 2]))
+
+        A = np.array(
+            [
+                [-6, 2 + 1j, -3, -5 + 2j],
+                [2 - 1j, 0, 2 - 1j, -5 + 4j],
+                [-3, 2 + 1j, 0, -4 + 3j],
+                [-5 - 2j, -5 - 4j, -4 - 3j, -6],
+            ]
+        )
+
+        with monkeypatch.context() as m:
+            m.setattr("numpy.random.choice", lambda x, y, p: (x, p))
+            s1, p = dev.sample(qml.PauliZ(0) @ qml.Hermitian(A, wires=[1, 2]))
+
+        # s1 should only contain the eigenvalues of
+        # the hermitian matrix tensor product Z
+        Z = np.diag([1, -1])
+        eigvals = np.linalg.eigvalsh(np.kron(Z, A))
+        assert set(np.round(s1, 8)).issubset(set(np.round(eigvals, 8)))
+
+        mean = s1 @ p
+        expected = 0.5 * (
+            -6 * np.cos(theta) * (np.cos(varphi) + 1)
+            - 2 * np.sin(varphi) * (np.cos(theta) + np.sin(phi) - 2 * np.cos(phi))
+            + 3 * np.cos(varphi) * np.sin(phi)
+            + np.sin(phi)
+        )
+        assert np.allclose(mean, expected, atol=tol, rtol=0)
+
+        var = (s1 ** 2) @ p - (s1 @ p).real ** 2
+        expected = (
+            1057
+            - np.cos(2 * phi)
+            + 12 * (27 + np.cos(2 * phi)) * np.cos(varphi)
+            - 2 * np.cos(2 * varphi) * np.sin(phi) * (16 * np.cos(phi) + 21 * np.sin(phi))
+            + 16 * np.sin(2 * phi)
+            - 8 * (-17 + np.cos(2 * phi) + 2 * np.sin(2 * phi)) * np.sin(varphi)
+            - 8 * np.cos(2 * theta) * (3 + 3 * np.cos(varphi) + np.sin(varphi)) ** 2
+            - 24 * np.cos(phi) * (np.cos(phi) + 2 * np.sin(phi)) * np.sin(2 * varphi)
+            - 8
+            * np.cos(theta)
+            * (
+                4
+                * np.cos(phi)
+                * (
+                    4
+                    + 8 * np.cos(varphi)
+                    + np.cos(2 * varphi)
+                    - (1 + 6 * np.cos(varphi)) * np.sin(varphi)
+                )
+                + np.sin(phi)
+                * (
+                    15
+                    + 8 * np.cos(varphi)
+                    - 11 * np.cos(2 * varphi)
+                    + 42 * np.sin(varphi)
+                    + 3 * np.sin(2 * varphi)
+                )
+            )
+        ) / 16
+        assert np.allclose(var, expected, atol=tol, rtol=0)
