--- tests/test_qubit_device.py
@@ -1,172 +1,0 @@
-# Copyright 2020 Xanadu Quantum Technologies Inc.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-Unit tests for the :mod:`pennylane.plugin.DefaultQubit` device.
-"""
-import cmath
-# pylint: disable=protected-access,cell-var-from-loop
-import math
-
-import pytest
-import pennylane as qml
-from pennylane import numpy as np, DeviceError
-from pennylane.operation import Operation
-from pennylane.plugins.default_qubit import (CRot3, CRotx, CRoty, CRotz,
-                                             Rot3, Rotx, Roty, Rotz,
-                                             Rphi, Y, Z, hermitian,
-                                             spectral_decomposition, unitary)
-
-THETA = np.linspace(0.11, 1, 3)
-PHI = np.linspace(0.32, 1, 3)
-VARPHI = np.linspace(0.02, 1, 3)
-
-@pytest.mark.parametrize("theta, phi, varphi", list(zip(THETA, PHI, VARPHI)))
-class TestTensorSample:
-    """Test tensor expectation values"""
-
-    def test_paulix_pauliy(self, theta, phi, varphi, monkeypatch, tol):
-        """Test that a tensor product involving PauliX and PauliY works correctly"""
-        dev = qml.device("default.qubit", wires=3, shots=10000)
-        dev.reset()
-        dev.apply(qml.RX(theta, wires=[0]))
-        dev.apply(qml.RX(phi, wires=[1]))
-        dev.apply(qml.RX(varphi, wires=[2]))
-        dev.apply(qml.CNOT(wires=[0, 1]))
-        dev.apply(qml.CNOT(wires=[1, 2]))
-
-        with monkeypatch.context() as m:
-            m.setattr("numpy.random.choice", lambda x, y, p: (x[0], p))
-            s1, p = dev.sample_basis_states(qml.PauliX(0) @ qml.PauliY(2))
-
-
-        print(s1)
-        print(p)
-        # s1 should only contain 1 and -1
-        assert np.allclose(s1 ** 2, 1, atol=tol, rtol=0)
-
-        mean = s1 @ p
-        expected = np.sin(theta) * np.sin(phi) * np.sin(varphi)
-        assert np.allclose(mean, expected, atol=tol, rtol=0)
-
-        var = (s1 ** 2) @ p - (s1 @ p).real ** 2
-        expected = (
-            8 * np.sin(theta) ** 2 * np.cos(2 * varphi) * np.sin(phi) ** 2
-            - np.cos(2 * (theta - phi))
-            - np.cos(2 * (theta + phi))
-            + 2 * np.cos(2 * theta)
-            + 2 * np.cos(2 * phi)
-            + 14
-        ) / 16
-        assert np.allclose(var, expected, atol=tol, rtol=0)
-
-    def test_pauliz_hadamard(self, theta, phi, varphi, monkeypatch, tol):
-        """Test that a tensor product involving PauliZ and PauliY and hadamard works correctly"""
-        dev = qml.device("default.qubit", wires=3)
-        dev.reset()
-        dev.apply(qml.RX(theta, wires=[0]))
-        dev.apply(qml.RX(phi, wires=[1]))
-        dev.apply(qml.RX(varphi, wires=[2]))
-        dev.apply(qml.CNOT(wires=[0, 1]))
-        dev.apply(qml.CNOT(wires=[1, 2]))
-
-        with monkeypatch.context() as m:
-            m.setattr("numpy.random.choice", lambda x, y, p: (x, p))
-            s1, p = dev.sample(qml.PauliZ(0) @ qml.Hadamard(1) @ qml.PauliY(2))
-
-        # s1 should only contain 1 and -1
-        assert np.allclose(s1 ** 2, 1, atol=tol, rtol=0)
-
-        mean = s1 @ p
-        expected = -(np.cos(varphi) * np.sin(phi) + np.sin(varphi) * np.cos(theta)) / np.sqrt(2)
-        assert np.allclose(mean, expected, atol=tol, rtol=0)
-
-        var = (s1 ** 2) @ p - (s1 @ p).real ** 2
-        expected = (
-            3
-            + np.cos(2 * phi) * np.cos(varphi) ** 2
-            - np.cos(2 * theta) * np.sin(varphi) ** 2
-            - 2 * np.cos(theta) * np.sin(phi) * np.sin(2 * varphi)
-        ) / 4
-        assert np.allclose(var, expected, atol=tol, rtol=0)
-
-    def test_hermitian(self, theta, phi, varphi, monkeypatch, tol):
-        """Test that a tensor product involving qml.Hermitian works correctly"""
-        dev = qml.device("default.qubit", wires=3)
-        dev.reset()
-        dev.apply(qml.RX(theta, wires=[0]))
-        dev.apply(qml.RX(phi, wires=[1]))
-        dev.apply(qml.RX(varphi, wires=[2]))
-        dev.apply(qml.CNOT(wires=[0, 1]))
-        dev.apply(qml.CNOT(wires=[1, 2]))
-
-        A = np.array(
-            [
-                [-6, 2 + 1j, -3, -5 + 2j],
-                [2 - 1j, 0, 2 - 1j, -5 + 4j],
-                [-3, 2 + 1j, 0, -4 + 3j],
-                [-5 - 2j, -5 - 4j, -4 - 3j, -6],
-            ]
-        )
-
-        with monkeypatch.context() as m:
-            m.setattr("numpy.random.choice", lambda x, y, p: (x, p))
-            s1, p = dev.sample(qml.PauliZ(0) @ qml.Hermitian(A, wires=[1, 2]))
-
-        # s1 should only contain the eigenvalues of
-        # the hermitian matrix tensor product Z
-        Z = np.diag([1, -1])
-        eigvals = np.linalg.eigvalsh(np.kron(Z, A))
-        assert set(np.round(s1, 8)).issubset(set(np.round(eigvals, 8)))
-
-        mean = s1 @ p
-        expected = 0.5 * (
-            -6 * np.cos(theta) * (np.cos(varphi) + 1)
-            - 2 * np.sin(varphi) * (np.cos(theta) + np.sin(phi) - 2 * np.cos(phi))
-            + 3 * np.cos(varphi) * np.sin(phi)
-            + np.sin(phi)
-        )
-        assert np.allclose(mean, expected, atol=tol, rtol=0)
-
-        var = (s1 ** 2) @ p - (s1 @ p).real ** 2
-        expected = (
-            1057
-            - np.cos(2 * phi)
-            + 12 * (27 + np.cos(2 * phi)) * np.cos(varphi)
-            - 2 * np.cos(2 * varphi) * np.sin(phi) * (16 * np.cos(phi) + 21 * np.sin(phi))
-            + 16 * np.sin(2 * phi)
-            - 8 * (-17 + np.cos(2 * phi) + 2 * np.sin(2 * phi)) * np.sin(varphi)
-            - 8 * np.cos(2 * theta) * (3 + 3 * np.cos(varphi) + np.sin(varphi)) ** 2
-            - 24 * np.cos(phi) * (np.cos(phi) + 2 * np.sin(phi)) * np.sin(2 * varphi)
-            - 8
-            * np.cos(theta)
-            * (
-                4
-                * np.cos(phi)
-                * (
-                    4
-                    + 8 * np.cos(varphi)
-                    + np.cos(2 * varphi)
-                    - (1 + 6 * np.cos(varphi)) * np.sin(varphi)
-                )
-                + np.sin(phi)
-                * (
-                    15
-                    + 8 * np.cos(varphi)
-                    - 11 * np.cos(2 * varphi)
-                    + 42 * np.sin(varphi)
-                    + 3 * np.sin(2 * varphi)
-                )
-            )
-        ) / 16
-        assert np.allclose(var, expected, atol=tol, rtol=0)
